+++ ksql-metastore/src/test/java/io/confluent/ksql/util/MetaStoreFixture.java
@@ -33,7 +33,6 @@ import io.confluent.ksql.serde.FormatInfo;
 import io.confluent.ksql.serde.KeyFormat;
 import io.confluent.ksql.serde.SerdeOption;
 import io.confluent.ksql.serde.ValueFormat;
-import io.confluent.ksql.util.timestamp.MetadataTimestampExtractionPolicy;
 import java.util.Optional;
 
 @SuppressWarnings("OptionalGetWithoutIsPresent")
