+++ ksql-streams/src/test/java/io/confluent/ksql/execution/streams/timestamp/MetadataTimestampExtractionPolicyTest.java
@@ -13,22 +13,23 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.util.timestamp;
+package io.confluent.ksql.execution.streams.timestamp;
 
 import com.google.common.testing.EqualsTester;
-import io.confluent.ksql.name.ColumnName;
-import io.confluent.ksql.schema.ksql.ColumnRef;
+import org.apache.kafka.streams.processor.UsePreviousTimeOnInvalidTimestamp;
 import org.junit.Test;
 
-public class StringTimestampExtractionPolicyTest {
+public class MetadataTimestampExtractionPolicyTest {
   @Test
   public void shouldTestEqualityCorrectly() {
     new EqualsTester()
         .addEqualityGroup(
-            new StringTimestampExtractionPolicy(ColumnRef.withoutSource(ColumnName.of("field1")), "yyMMddHHmmssZ"),
-            new StringTimestampExtractionPolicy(ColumnRef.withoutSource(ColumnName.of("field1")), "yyMMddHHmmssZ"))
-        .addEqualityGroup(new StringTimestampExtractionPolicy(ColumnRef.withoutSource(ColumnName.of("field2")), "yyMMddHHmmssZ"))
-        .addEqualityGroup(new StringTimestampExtractionPolicy(ColumnRef.withoutSource(ColumnName.of("field1")), "ddMMyyHHmmssZ"))
+            new MetadataTimestampExtractionPolicy(),
+            new MetadataTimestampExtractionPolicy())
+        .addEqualityGroup(
+            new MetadataTimestampExtractionPolicy(new UsePreviousTimeOnInvalidTimestamp()),
+            new MetadataTimestampExtractionPolicy(new UsePreviousTimeOnInvalidTimestamp())
+        )
         .testEquals();
   }
 }
